Add relative linking support
